+++ src/main/java/com/ivanvinski/kunvertuh/mvp/presenter/ConverterPresenter.java
@@ -22,33 +22,38 @@ import com.google.common.eventbus.Subscribe;
 import com.ivanvinski.kunvertuh.event.EventStream;
 import com.ivanvinski.kunvertuh.i18n.Language;
 import com.ivanvinski.kunvertuh.measurement.Unit;
-import com.ivanvinski.kunvertuh.measurement.UnitConverter;
+import com.ivanvinski.kunvertuh.mvp.model.ConverterModel;
 import com.ivanvinski.kunvertuh.mvp.view.ConverterView;
 import com.ivanvinski.kunvertuh.util.DoubleStringConverter;
 import java.text.NumberFormat;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 public class ConverterPresenter<U extends Unit> extends
-    Presenter<ConverterView<U>, UnitConverter<U>> {
+    Presenter<ConverterView<U>, ConverterModel<U>> {
 
   private DoubleStringConverter valueConverter = new DoubleStringConverter();
 
-  public ConverterPresenter(ConverterView<U> view, UnitConverter<U> model,
+  public ConverterPresenter(ConverterView<U> view, ConverterModel<U> model,
       EventStream eventStream) {
     super(view, model, eventStream);
   }
 
   @Override
   public void onInitialized() {
-    for (U unit : getModel().getSupportedUnits()) {
+    getModel().getCategories()
+        .stream()
+        .map(getModel()::getUnits)
+        .forEach(this::divideUnitsAndAppendUnitsCard);
+    for (U unit : getModel().getAllUnits()) {
       getView().setOnConversionAction(unit, value -> onConversionRequested(unit, value));
     }
   }
 
   @Override
   public void onLanguageChanged(Language language) {
-    for (U unit : getModel().getSupportedUnits()) {
+    for (U unit : getModel().getAllUnits()) {
       getView().setUnitPrompt(unit, language.getString(unit.toString()));
     }
   }
@@ -56,20 +61,28 @@ public class ConverterPresenter<U extends Unit> extends
   @Subscribe
   public void onNumberFormatChanged(NumberFormat numberFormat) {
     Map<U, Double> tmpValues = new HashMap<>();
-    for (U unit : getModel().getSupportedUnits()) {
+    for (U unit : getModel().getAllUnits()) {
       valueConverter.parse(getView().getUnitValue(unit));
       tmpValues.put(unit, valueConverter.parse(getView().getUnitValue(unit)));
     }
     valueConverter.setValueFormat(numberFormat);
-    for (U unit : getModel().getSupportedUnits()) {
+    for (U unit : getModel().getAllUnits()) {
       getView().setUnitValue(unit, valueConverter.format(tmpValues.get(unit)));
     }
   }
 
   public void onConversionRequested(U unit, String value) {
-    getModel().convert(valueConverter.parse(value), unit);
-    for (U otherUnit : getModel().getSupportedUnits()) {
-      getView().setUnitValue(otherUnit, valueConverter.format(getModel().getValue(otherUnit)));
+    getModel().getConverter().convert(valueConverter.parse(value), unit);
+    for (U otherUnit : getModel().getAllUnits()) {
+      String formattedValue = valueConverter.format(getModel().getConverter().getValue(otherUnit));
+      getView().setUnitValue(otherUnit, formattedValue);
     }
+  }
+
+  private void divideUnitsAndAppendUnitsCard(List<U> units) {
+    int halfTheSize = (int) Math.ceil((float) units.size() / 2.0);
+    List<U> leftColumn = units.subList(0, halfTheSize);
+    List<U> rightColumn = units.subList(halfTheSize, units.size());
+    getView().appendUnitsCard(leftColumn, rightColumn);
   }
 }